コードの改善と修正
