+++ app/src/main/java/com/saradabar/cpadcustomizetool/view/flagment/WelcomeFragment.java
@@ -22,7 +22,7 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.widget.AppCompatTextView;
 import androidx.fragment.app.Fragment;
 
-import com.saradabar.cpadcustomizetool.MainActivity;
+import com.saradabar.cpadcustomizetool.view.activity.CheckActivity;
 import com.saradabar.cpadcustomizetool.R;
 import com.saradabar.cpadcustomizetool.util.Constants;
 import com.saradabar.cpadcustomizetool.util.Preferences;
@@ -39,7 +39,7 @@ public class WelcomeFragment extends Fragment implements WelcomePage.OnChangeLis
         view.findViewById(R.id.wel_no).setOnClickListener(v -> requireActivity().finishAffinity());
         view.findViewById(R.id.wel_yes).setOnClickListener(v -> {
             Preferences.save(requireActivity(), Constants.KEY_FLAG_APP_WELCOME_COMPLETE, true);
-            startActivity(new Intent(requireActivity(), MainActivity.class).addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION));
+            startActivity(new Intent(requireActivity(), CheckActivity.class).addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION));
             requireActivity().overridePendingTransition(0, 0);
             requireActivity().finish();
         });
